pyproject.toml
[build-system]
requires = ["maturin>=1.9,<2.0"]
build-backend = "maturin"

[project]
name = "inpaint"
requires-python = ">=3.8"
classifiers = [
    'License :: OSI Approved :: European Union Public Licence 1.2 (EUPL 1.2)',
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Programming Language :: Python :: 3.14',
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Multimedia :: Graphics",
]
readme = "README.md"
license = "EUPL-1.2"
authors = [{ name = "Gilles Vink", email = "gillesvink@noreply.codeberg.org" }]
maintainers = [
    { name = "Gilles Vink", email = "gillesvink@noreply.codeberg.org" },
]
description = "Easy to use library for inpainting images."
dynamic = ["version"]
dependencies = [
    "numpy>=1.16.0",
    "pillow>=10.0.0",
]


[project.urls]
Homepage = "https://codeberg.org/gillesvink/inpaint"
"Bug Tracker" = "https://codeberg.org/gillesvink/inpaint/issues"
Changelog = "https://codeberg.org/gillesvink/inpaint/src/branch/main/CHANGELOG.md"

[tool.maturin]
python-source = "python"
features = ["pyo3/extension-module", "python-bindings"]

[tool.uv]
cache-keys = [
    { file = "pyproject.toml" },
    { file = "Cargo.toml" },
    { file = "**/*.rs" },
]

[tool.uv.workspace]
members = [
    "examples/python",
]

[dependency-groups]
dev = ["pytest>=8.3.0"]
